Fehler behoben in rtl8139.c
git-svn-id: svn://svn.code.sf.net/p/prettyos/code/trunk/Source@144 31957799-4632-4615-9793-72db6d48bcac

+++ kernel/rtl8139.c
@@ -154,7 +154,7 @@ void rtl8139_handler(struct regs* r)
     for(int8_t i=16;i<18;i++) {printformat("%y ",network_buffer[i]);}
 
     printformat("\n");
-    for(int8_t i=18;i<=length;i++) {printformat("%y ",network_buffer[i]);}
+    for(int32_t i=18;i<=length;i++) {printformat("%y ",network_buffer[i]);}
     printformat("\n--------------------------------------------------------------------------------\n");
 
     settextcolor(15,0);
